Able to save file every save_interval

+++ util.py
@@ -136,7 +136,7 @@ def to_Variable(*arrays, **kwargs):
         if isinstance(array, int):
             array = [array]
         if isinstance(array, np.ndarray) or isinstance(array, list):
-            array = torch.tensor(array)
+            array = torch.tensor(array).float()
         if isinstance(array, torch.FloatTensor) or isinstance(array, torch.LongTensor) or isinstance(array, torch.ByteTensor):
             array = Variable(array, requires_grad = requires_grad)
         if is_cuda:
